encapsulate rest of cmacc csc2 code
