+++ db/CMakeLists.txt
@@ -101,6 +101,7 @@ set(src
   dohast.c
   machclass.c
   osqluprec.c
+  macc_glue.c
   ${PROJECT_BINARY_DIR}/protobuf/bpfunc.pb-c.c
   ${PROJECT_SOURCE_DIR}/tools/cdb2_dump/cdb2_dump.c
   ${PROJECT_SOURCE_DIR}/tools/cdb2_load/cdb2_load.c
