+++ schemachange/sc_csc2.c
@@ -21,6 +21,7 @@
 #include "sc_csc2.h"
 #include "debug_switches.h"
 #include "logmsg.h"
+#include "macc_glue.h"
 
 static int schema_cmp(struct dbenv *dbenv, struct dbtable *db,
                       const char *csc2cmp);
@@ -71,9 +72,7 @@ int check_table_schema(struct dbenv *dbenv, const char *table,
     if (meta_csc2) {
         /* see if the loaded schema differs from the schema contained
          * in our meta table. */
-        dyns_init_globals();
         rc = schema_cmp(dbenv, db, meta_csc2);
-        dyns_cleanup_globals();
         if (rc != 0) {
             logmsg(LOGMSG_ERROR, "SCHEMA MIS-MATCH FOR TABLE %s.\n", table);
             logmsg(LOGMSG_ERROR, "THIS IS MY SCHEMA (VERSION %d):-\n", version);
@@ -107,19 +106,12 @@ int check_table_schema(struct dbenv *dbenv, const char *table,
 static int schema_cmp(struct dbenv *dbenv, struct dbtable *db,
                       const char *csc2cmp)
 {
-    int rc =
-        dyns_load_schema_string((char *)csc2cmp, dbenv->envname, db->tablename);
-    if (rc) {
-        logmsg(LOGMSG_ERROR, "schema_cmp: error loading comparison schema\n");
-        return -1;
-    }
+    struct errstat err = {0};
+    int rc;
 
-    /* create the .NEW. tags, but don't update db->lrl, numix, numblobs
-     * or anything like that. */
-    rc = add_cmacc_stmt_no_side_effects(db, 1);
+    rc = populate_db_with_alt_schema(dbenv, db, (char *)csc2cmp, &err);
     if (rc) {
-        logmsg(LOGMSG_ERROR,
-               "schema_cmp: error creating schema from comparison text\n");
+        logmsg(LOGMSG_ERROR, "%s\ncsc2: \"%s\"\n", err.errstr, csc2cmp);
         return -1;
     }
